Added the changes

+++ src/app/components/home/home.component.ts
@@ -1,21 +1,19 @@
-import { Component , ViewContainerRef} from '@angular/core';
+import { Component,OnInit} from '@angular/core';
 import { FormGroup, FormBuilder, Validators } from '@angular/forms';
 import { Router, ActivatedRoute } from '@angular/router';
 import { environment } from '../../../environments/environment';
 import { AppService } from '../../app.service';
-import { AlertService } from '../../shared/alerts/alerts.service';
-import {ToastsManager, Toast} from 'ng2-toastr/ng2-toastr';
 import { CookieService } from 'angular2-cookie';
+import { NgxSpinnerService } from 'ngx-spinner';
 
 declare var $: any; 
 @Component({
   
   selector: 'home',
   templateUrl: './home.component.html',
-  styleUrls: ['./home.component.css'],
-  providers:[AppService,AlertService,CookieService]
+  providers:[AppService,CookieService,NgxSpinnerService]
 })
-export class HomeComponent {
+export class HomeComponent implements OnInit {
          
   loginform: any;
   customerId:any;
@@ -28,34 +26,21 @@ export class HomeComponent {
   emailPattern = "^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-z]{2,4}$"; 
   password:any;
   userId:any;
-  constructor(  private toastr:ToastsManager,private _vcr: ViewContainerRef,private route: ActivatedRoute,private _cookieService: CookieService,
-      private fb: FormBuilder, private router: Router,private appService: AppService,private alertService : AlertService) {
-        this.route.params.subscribe(params => {
-       
-          // else if (params['Uid']>0) {
-          //   this.ActivatetheUser(params['Uid']);
-          // }
-        });
-        //this.toastr.setRootViewContainerRef(_vcr);
+  constructor(  private spinner: NgxSpinnerService,private route: ActivatedRoute,private _cookieService: CookieService,
+      private fb: FormBuilder, private router: Router,private appService: AppService) {
+        this.spinner.show();
         this.tkeyres  = this._cookieService.get('token');
         if(this.tkeyres !=null)
         {
           this.GetLogin(this.tkeyres);
         }
+        else
+        {
+          window.location.href = 'http://demo.tenendus.com:1070/customerLogin';
+        }
   }
 
-//   login1(username: string, password: string) {
-//     return this.http.post<any>('/api/authenticate', { username: username, password: password })
-//         .map(user => {
-//             // login successful if there's a jwt token in the response
-//             if (user && user.token) {
-//                 // store user details and jwt token in local storage to keep user logged in between page refreshes
-//                 localStorage.setItem('currentUser', JSON.stringify(user));
-//             }
 
-//             return user;
-//         });
-// }
 
 
   GetLogin(res)
@@ -63,10 +48,20 @@ export class HomeComponent {
     this.tkey.UserToken = res;
     this.appService.GetCustomerToken(this.tkey).subscribe(
       data => {
+        if(data.UserId>0 && data.IsActive == true)
+        {
+        this.spinner.hide();
         sessionStorage.setItem('isLoggedin', JSON.stringify('true'));
         sessionStorage.setItem('userData', JSON.stringify(data));
         this.router.navigateByUrl('app-dashboardview');
+        this.tkeyres = null;
+        this.tkey.UserToken= null;
         this._cookieService.remove('token'); 
+        }
+        else
+        {
+          window.location.href = 'http://demo.tenendus.com:1070/customerLogin';
+        }
       })
   }
 
@@ -75,7 +70,7 @@ export class HomeComponent {
 
 
   ngOnInit() {
-  
+    this.spinner.show();
   }
 }
